+++ src/ResponseBuilders/GuzzleResponseBuilder.php
@@ -5,13 +5,21 @@ namespace EasyHttp\MockBuilder\ResponseBuilders;
 use EasyHttp\MockBuilder\Contracts\ResponseBuilder;
 use GuzzleHttp\Promise\FulfilledPromise;
 use GuzzleHttp\Psr7\Response;
+use Psr\Http\Message\RequestInterface;
 
 class GuzzleResponseBuilder implements ResponseBuilder
 {
     private int $statusCode = 200;
     private array $headers = [];
     private string $body = '';
+    private $bodyHandler;
     private array $json = [];
+    private ?RequestInterface $request;
+
+    public function setRequest(?RequestInterface $request = null)
+    {
+        $this->request = $request;
+    }
 
     public function statusCode(int $statusCode): self
     {
@@ -42,10 +50,19 @@ class GuzzleResponseBuilder implements ResponseBuilder
         return $this->body;
     }
 
-    public function body(string $body): self
+    public function body($body): self
     {
-        $this->body = $body;
         $this->json = [];
+        $this->body = '';
+        $this->bodyHandler = null;
+
+        if (is_callable($body)) {
+            $this->bodyHandler = $body;
+        }
+
+        if (is_string($body)) {
+            $this->body = $body;
+        }
 
         return $this;
     }
@@ -65,7 +82,16 @@ class GuzzleResponseBuilder implements ResponseBuilder
 
     public function response(): FulfilledPromise
     {
-        $body = $this->json ? json_encode($this->json, true) : $this->body;
+        $body = '';
+
+        if ($this->bodyHandler) {
+            $clousure = $this->bodyHandler;
+            $body = $clousure($this->request);
+        } elseif (!empty($this->json)) {
+            $body = json_encode($this->json, true);
+        } elseif (!empty($this->body)) {
+            $body = $this->body;
+        }
 
         return new FulfilledPromise(
             new Response($this->statusCode, $this->headers, $body)